Image Segmantation With CNN
In this study, using a dataset of runway images from the Microsoft Flight Simulator game, a Convolutional Neural Network (CNN) algorithm is used to segment runway images.

The dataset consists of realistic runway images collected from the simulator. The images are preprocessed for training and evaluation of the deep learning model.

In this study, a CNN-based image segmentation model is developed. The model is trained to accurately discriminate runway regions. The performance of the model is tested by using appropriate metrics.

The results show that the proposed CNN model shows high performance in segmentation of runway images. This study provides a powerful image processing solution that can be used in flight simulation and aviation applications.

In future studies, different architectures and algorithms can be tested, more comprehensive datasets can be used, and model performance can be improved. Thus, more advanced systems can be created that can be used in critical aviation tasks such as runway detection and flight planning.